emulator/requirements_emulator.txt
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==4.0.2
attrs==23.1.0
causalpaca==1.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
docker-pycreds==0.4.0
filelock==3.12.0
frozenlist==1.3.3
fsspec==2023.5.0
gitdb==4.0.10
GitPython==3.1.31
huggingface-hub==0.14.1
hydra-core==1.3.2
idna==3.4
Jinja2==3.1.2
lightning-utilities==0.8.0
lit==16.0.5
MarkupSafe==2.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.3.0
packaging==23.1
pathtools==0.1.2
Pillow==9.5.0
protobuf==4.23.1
psutil==5.9.5
pytorch-lightning==2.0.2
PyYAML==6.0
requests==2.31.0
safetensors==0.3.1
sentry-sdk==1.24.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sympy==1.12
timm==0.9.2
torch==2.0.1
torchmetrics==0.11.4
torchvision==0.15.2
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.6.2
urllib3==1.26.16
wandb==0.15.3
yarl==1.9.2
